feat: Implement team management feature in settings
- Added Team component to manage team members, including invite functionality.
- Integrated team management into the settings page with user permissions check.
- Created a new layout for map pages to enhance user experience.
- Developed a comprehensive map page with sidebar and controls for air quality data visualization.
- Established a dynamic layout system for dashboard, auth, and map routes with centralized configuration.
- Documented layout system and configurations for better maintainability and clarity.
