+++ src/platform/src/app/(auth)/account/creation/get-started/page.jsx
