+++ src/platform/src/app/(auth)/account/creation/individual/interest/[id]/[token]/page.jsx
