+++ src/platform/src/app/(auth)/account/creation/individual/register/page.jsx
