+++ src/platform/src/app/(auth)/account/creation/individual/verify-email/page.jsx
