+++ src/platform/src/app/(auth)/account/creation/organisation/verify/[id]/create-org/details/page.jsx
