+++ src/platform/src/app/(auth)/account/creation/page.jsx
