+++ src/platform/src/app/(auth)/account/forgotPwd/page.jsx
