+++ src/platform/src/app/(auth)/account/forgotPwd/reset/page.jsx
