+++ src/platform/src/app/(auth)/account/login/page.jsx
