+++ src/platform/src/app/(auth)/layout.jsx
@@ -1,0 +1,5 @@
+import AuthLayout from '@/layouts/AuthLayout';
+
+export default function AuthRouteGroupLayout({ children }) {
+  return <AuthLayout>{children}</AuthLayout>;
+}
