+++ src/platform/src/app/(pages)/(layout-1)/analytics/page.jsx
@@ -2,7 +2,6 @@
 
 import React, { useEffect, useState, useRef } from 'react';
 import withAuth from '@/core/utils/protectedRoute';
-import Layout from '@/components/Layout';
 import { useDispatch, useSelector } from 'react-redux';
 import OverView from './view/OverView';
 import AlertBox from '@/components/AlertBox';
@@ -40,7 +39,7 @@ const AuthenticatedHomePage = () => {
   }, [dispatch, preferenceData]);
 
   return (
-    <Layout topbarTitle={'Analytics'} noBorderBottom pageTitle={'Analytics'}>
+    <>
       <AlertBox
         type={alert.type}
         message={alert.message}
@@ -48,7 +47,7 @@ const AuthenticatedHomePage = () => {
         hide={() => setAlert({ ...alert, show: false })}
       />
       <OverView />
-    </Layout>
+    </>
   );
 };
