+++ src/platform/src/app/(pages)/(layout-1)/collocation/collocate_success/page.jsx
