+++ src/platform/src/app/(pages)/(layout-1)/collocation/reports/[device]/page.jsx
@@ -1,6 +1,5 @@
 'use client';
 
-import Layout from '@/components/Layout';
 import NavigationBreadCrumb from '@/components/Navigation/breadcrumb';
 import Box from '@/components/Collocation/Report/box';
 import PollutantDropdown from '@/components/Collocation/Report/PollutantDropdown';
@@ -99,85 +98,83 @@ const Reports = () => {
   );
 
   return (
-    <Layout>
-      <div className="flex flex-col gap-6">
-        <NavigationBreadCrumb breadcrumbData={breadcrumbData} />
+    <div className="flex flex-col gap-6">
+      <NavigationBreadCrumb breadcrumbData={breadcrumbData} />
 
-        {isLoading ? (
-          <div className="flex justify-center items-center py-20">
-            <Spinner width={40} height={40} />
+      {isLoading ? (
+        <div className="flex justify-center items-center py-20">
+          <Spinner width={40} height={40} />
+        </div>
+      ) : (
+        <>
+          <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
+            <Box
+              title="Intra Sensor Correlation"
+              subtitle="Comparison of data between sensors within the same device."
+            />
+            <Box
+              title="Inter Sensor Correlation"
+              subtitle="Comparison of data between sensors across different devices."
+            />
           </div>
-        ) : (
-          <>
-            <div className="grid grid-cols-1 lg:grid-cols-2 gap-6">
-              <Box
-                title="Intra Sensor Correlation"
-                subtitle="Comparison of data between sensors within the same device."
-              />
-              <Box
-                title="Inter Sensor Correlation"
-                subtitle="Comparison of data between sensors across different devices."
+
+          <ContentBox>
+            <div className="flex justify-between items-center mb-6">
+              <h2 className="text-xl font-semibold text-gray-900">
+                Device Report: {device}
+              </h2>
+              <PollutantDropdown
+                selectedPollutant={pmConcentration}
+                onPollutantChange={handlePollutantChange}
               />
             </div>
 
-            <ContentBox>
-              <div className="flex justify-between items-center mb-6">
-                <h2 className="text-xl font-semibold text-gray-900">
-                  Device Report: {device}
-                </h2>
-                <PollutantDropdown
-                  selectedPollutant={pmConcentration}
-                  onPollutantChange={handlePollutantChange}
-                />
-              </div>
-
-              {!isEmpty(filteredStatistics) ? (
-                <>
-                  <div className="mb-8">
-                    <CorrelationChart
-                      data={filteredStatistics}
-                      pollutant={`pm${pmConcentration}`}
-                    />
-                  </div>
-
-                  <div className="mb-6">
-                    <CustomLegend data={filteredStatistics} />
-                  </div>
-
-                  <CustomTable
+            {!isEmpty(filteredStatistics) ? (
+              <>
+                <div className="mb-8">
+                  <CorrelationChart
                     data={filteredStatistics}
-                    columns={[
-                      { key: 'batch_id', label: 'Batch ID' },
-                      { key: 'correlation_coefficient', label: 'Correlation' },
-                      { key: 'r_squared', label: 'R²' },
-                      { key: 'rmse', label: 'RMSE' },
-                      { key: 'mean_bias', label: 'Mean Bias' },
-                    ]}
+                    pollutant={`pm${pmConcentration}`}
                   />
-                </>
-              ) : (
-                <div className="text-center py-12 text-gray-500">
-                  <p>No correlation data available for PM{pmConcentration}</p>
-                  <p className="text-sm mt-2">
-                    Try selecting a different pollutant or check if data exists
-                    for this device.
-                  </p>
                 </div>
-              )}
-            </ContentBox>
-          </>
-        )}
 
-        {error && (
-          <Toast
-            type="error"
-            timeout={5000}
-            message={error}
-            clearData={() => setError(null)}
-          />
-        )}
-      </div>
-    </Layout>
+                <div className="mb-6">
+                  <CustomLegend data={filteredStatistics} />
+                </div>
+
+                <CustomTable
+                  data={filteredStatistics}
+                  columns={[
+                    { key: 'batch_id', label: 'Batch ID' },
+                    { key: 'correlation_coefficient', label: 'Correlation' },
+                    { key: 'r_squared', label: 'R²' },
+                    { key: 'rmse', label: 'RMSE' },
+                    { key: 'mean_bias', label: 'Mean Bias' },
+                  ]}
+                />
+              </>
+            ) : (
+              <div className="text-center py-12 text-gray-500">
+                <p>No correlation data available for PM{pmConcentration}</p>
+                <p className="text-sm mt-2">
+                  Try selecting a different pollutant or check if data exists
+                  for this device.
+                </p>
+              </div>
+            )}
+          </ContentBox>
+        </>
+      )}
+
+      {error && (
+        <Toast
+          type="error"
+          timeout={5000}
+          message={error}
+          clearData={() => setError(null)}
+        />
+      )}
+    </div>
   );
 };
