+++ src/platform/src/app/(pages)/(layout-1)/layout.jsx
@@ -1,0 +1,5 @@
+import PagesLayout from '@/layouts/PagesLayout';
+
+export default function PagesRouteGroupLayout({ children }) {
+  return <PagesLayout>{children}</PagesLayout>;
+}
