+++ src/platform/src/app/(pages)/(layout-1)/settings/Tabs/Team.jsx
