+++ src/platform/src/app/(pages)/(layout-2)/layout.jsx
@@ -1,0 +1,5 @@
+import MapLayout from '@/layouts/MapLayout';
+
+export default function MapRouteLayout({ children }) {
+  return <MapLayout>{children}</MapLayout>;
+}
