+++ src/platform/src/common/features/theme-customizer/context/ThemeContext.jsx
@@ -1,3 +1,4 @@
+'use client';
 import React, {
   createContext,
   useState,
