+++ src/platform/src/common/layouts/PagesLayout.jsx
@@ -1,0 +1,90 @@
+'use client';
+
+import { usePathname } from 'next/navigation';
+import { useSelector } from 'react-redux';
+import { useRouter } from 'next/navigation';
+import Head from 'next/head';
+import AuthenticatedSideBar from '@/components/SideBar/AuthenticatedSidebar';
+import TopBar from '@/components/TopBar';
+import SideBarDrawer from '@/components/SideBar/SideBarDrawer';
+import MaintenanceBanner from '@/components/MaintenanceBanner';
+import useUserPreferences from '@/core/hooks/useUserPreferences';
+import useInactivityLogout from '@/core/hooks/useInactivityLogout';
+import useMaintenanceStatus from '@/core/hooks/useMaintenanceStatus';
+import { useGetActiveGroup } from '@/core/hooks/useGetActiveGroupId';
+import { useTheme } from '@/features/theme-customizer/hooks/useTheme';
+import { THEME_LAYOUT } from '@/features/theme-customizer/constants/themeConstants';
+import { LAYOUT_CONFIGS, DEFAULT_CONFIGS } from './layoutConfigs';
+
+/**
+ * Dashboard Layout Component
+ * Provides a shared layout for dashboard routes (Home, Analytics, Settings, Collocation)
+ * with dynamic configuration based on the current route
+ */
+export default function PagesLayout({ children }) {
+  const pathname = usePathname();
+  const router = useRouter();
+  const { userID } = useGetActiveGroup();
+  const isCollapsed = useSelector((state) => state.sidebar.isCollapsed);
+  const { maintenance } = useMaintenanceStatus();
+  const isMapPage = router.pathname === '/map';
+
+  // Get route configuration based on current pathname
+  const routeConfig =
+    LAYOUT_CONFIGS.DASHBOARD[pathname] || DEFAULT_CONFIGS.DASHBOARD;
+
+  // Initialize hooks
+  useUserPreferences();
+  useInactivityLogout(userID);
+
+  // Get current layout (compact or wide)
+  const { layout } = useTheme();
+
+  // Determine container classes based on layout preference
+  const containerClasses = !isMapPage
+    ? layout === THEME_LAYOUT.COMPACT
+      ? 'max-w-7xl mx-auto flex flex-col gap-8 px-4 py-4 md:px-6 lg:py-8 lg:px-8'
+      : 'w-full flex flex-col gap-8 px-4 py-4 md:px-6 lg:py-8 lg:px-8'
+    : '';
+
+  return (
+    <div className="flex overflow-hidden min-h-screen" data-testid="layout">
+      <Head>
+        <title>{routeConfig.pageTitle}</title>
+        <meta property="og:title" content={routeConfig.pageTitle} key="title" />
+      </Head>
+
+      {/* Sidebar */}
+      <aside className="fixed left-0 top-0 z-50 text-sidebar-text transition-all duration-300">
+        <AuthenticatedSideBar />
+      </aside>
+
+      {/* Main Content */}
+      <main
+        className={`flex-1 transition-all duration-300 
+          ${isMapPage ? 'overflow-hidden' : 'overflow-y-auto'} 
+          ${isCollapsed ? 'lg:ml-[88px]' : 'lg:ml-[256px]'}`}
+      >
+        <div className={`overflow-hidden ${containerClasses}`}>
+          {/* Maintenance Banner */}
+          {maintenance && <MaintenanceBanner maintenance={maintenance} />}
+
+          {/* TopBar */}
+          <TopBar
+            topbarTitle={routeConfig.topbarTitle}
+            noBorderBottom={routeConfig.noBorderBottom}
+            showSearch={routeConfig.showSearch}
+          />
+
+          {/* Content */}
+          <div className="text-text transition-all duration-300 overflow-hidden">
+            {children}
+          </div>
+        </div>
+      </main>
+
+      {/* SideBar Drawer */}
+      <SideBarDrawer />
+    </div>
+  );
+}
